requirements.txt
# Quiz App Dependencies
flask==3.0.0
werkzeug==3.0.1
psycopg2-binary==2.9.11
gunicorn==21.2.0
